feat: Result UI State 구현 . State 구현 . State UI에 부착

+++ app/src/main/java/com/sghore/needtalk/data/model/entity/FriendEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class FriendEntity(
     @PrimaryKey
     val userId: String,
-    val experiencePoint: Int,
+    val experiencePoint: Float,
     val friendshipPoint: Int,
     val createTime: Long = System.currentTimeMillis()
 )
