+++ app/src/main/java/com/sghore/needtalk/domain/model/UserData.kt
@@ -6,6 +6,6 @@ data class UserData(
     val userId: String,
     val name: String,
     val profileImage: ImageBitmap,
-    val experiencePoint: Int,
+    val experiencePoint: Float,
     val friendshipPoint: Int
 )
